requirements.txt
docopt==0.6.2
ecdsa==0.13
Fabric==1.10.2
paramiko==1.15.2
pycrypto==2.6.1
PyYAML==3.11
wheel==0.24.0
